# public/docs/openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Decent Event Management API",
        "description": "Event ticketing and management system API with registration, payment, ticketing, check-in, and reporting capabilities.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://api.example.com/v1",
            "description": "Production server"
        },
        {
            "url": "http://localhost:8000/api/v1",
            "description": "Local development server"
        }
    ],
    "paths": {
...
        "/admin/registrations": {
            "get": {
                "tags": [
                    "Registrations"
                ],
                "summary": "List registrations with search and filters",
                "operationId": "5de4262d9ce72fe8ccf80486c157d2cc",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Search by registration number, attendee full name, or attendee mobile",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Filter by registration status",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "ticket_type_id",
                        "in": "query",
                        "description": "Filter by ticket type id",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "date_from",
                        "in": "query",
                        "description": "Filter by created_at on or after this date",
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "date_to",
                        "in": "query",
                        "description": "Filter by created_at on or before this date",
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Results per page, capped at 100",
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paginated list of registrations",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "ulid": {
                                                        "type": "string"
                                                    },
                                                    "registration_number": {
                                                        "type": "string"
                                                    },
                                                    "status": {
                                                        "type": "string"
                                                    },
                                                    "participation_type": {
                                                        "type": "string"
                                                    },
                                                    "adults_count": {
                                                        "type": "integer"
                                                    },
                                                    "children_count": {
                                                        "type": "integer"
                                                    },
                                                    "subtotal_paisa": {
                                                        "type": "integer"
                                                    },
                                                    "discount_paisa": {
                                                        "type": "integer"
                                                    },
                                                    "total_paisa": {
                                                        "type": "integer"
                                                    },
                                                    "currency": {
                                                        "type": "string"
                                                    },
                                                    "discount_code": {
                                                        "type": "string",
                                                        "nullable": true
                                                    },
                                                    "comments": {
                                                        "type": "string",
                                                        "nullable": true
                                                    },
                                                    "special_notes": {
                                                        "type": "string",
                                                        "nullable": true
                                                    },
                                                    "source": {
                                                        "type": "string"
                                                    },
                                                    "submitted_at": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "nullable": true
                                                    },
                                                    "confirmed_at": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "nullable": true
                                                    },
                                                    "cancelled_at": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "nullable": true
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "format": "date-time"
                                                    },
                                                    "attendee": {
                                                        "type": "object"
                                                    },
                                                    "ticket_type": {
                                                        "type": "object"
                                                    },
                                                    "guests": {
                                                        "type": "array",
                                                        "items": {
                                                            "type": "object"
                                                        }
                                                    },
                                                    "event_session": {
                                                        "type": "object"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "links": {
                                            "type": "object"
                                        },
                                        "meta": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Missing registration.view_any permission"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Token issued by the relevant login endpoint. Admin, attendee, and scanner guards each mint tokens scoped to that guard and cannot be used interchangeably.",
                "bearerFormat": "Sanctum personal access token",
                "scheme": "bearer"
            }
        }
    },
...
}